requirements.txt
filterpy
numba
numpy
numpy
opencv-python
Pillow
scipy
torch
torchvision
tqdm
